control_loop module
