+++ src/server/controlchan/mod.rs
@@ -5,6 +5,9 @@ use command::Command;
 
 pub(crate) mod handler;
 
+pub(crate) mod control_loop;
+pub(crate) use control_loop::{spawn_control_channel_loop, ControlParams};
+
 pub(super) mod commands;
 
 mod parse_error;
